리액트앱키 gitignore 처리
